+++ src/processor/ControllerInterface.kt
@@ -1,7 +1,9 @@
 package processor
 
+import java.math.BigDecimal
+
 interface ControllerInterface {
     fun run()
-    fun matrixAddition(a: Array<Array<Int>>, b: Array<Array<Int>>): Array<Array<Int>>
-    fun matrixScalarMultiplication(a: Array<Array<Int>>, scalar: Int): Array<Array<Int>>
+    fun matrixAddition(a: Array<Array<BigDecimal>>, b: Array<Array<BigDecimal>>): Array<Array<BigDecimal>>
+    fun matrixScalarMultiplication(a: Array<Array<BigDecimal>>, scalar: BigDecimal): Array<Array<BigDecimal>>
 }
