+++ src/processor/ModelInterface.kt
@@ -1,6 +1,8 @@
 package processor
 
+import java.math.BigDecimal
+
 interface ModelInterface {
-    fun matrixAddition(a: Array<Array<Int>>, b: Array<Array<Int>>): Array<Array<Int>>
-    fun matrixScalarMultiplication(a: Array<Array<Int>>, scalar: Int): Array<Array<Int>>
+    fun matrixAddition(a: Array<Array<BigDecimal>>, b: Array<Array<BigDecimal>>): Array<Array<BigDecimal>>
+    fun matrixScalarMultiplication(a: Array<Array<BigDecimal>>, scalar: BigDecimal): Array<Array<BigDecimal>>
 }
